Annotate int type with actual mjt* type in mujoco.h
This PR annoated int type with the actual mjt* type in comment as
"parameter_name is mjtTYPE". This helps external parsers and document
generators to understand what the type MuJoCo refers to as in C, all
these are int.

The usefulness can be seen in swift-mujoco:

Before this PR:

https://github.com/liuliu/swift-mujoco/blob/0df02e2bc1eb07d74cb97f03fe1e24b7cd944428/Sources/MjrContext%2BFunctions.swift#L50

After this PR, this interface is generated:

https://github.com/liuliu/swift-mujoco/blob/main/Sources/MjrContext%2BFunctions.swift#L54

Next:

I need to figure out how to tag nullability for parameters. Would:
"parameter_name is nullable" acceptable one in comment?